project/requirements.txt
fastapi==0.116.1
httpx==0.28.1
joblib==1.5.2
matplotlib==3.10.5
numpy==2.3.2
pandas==2.3.3
pydantic==2.12.5
PyYAML==6.0.2
scikit-learn==1.7.2
seaborn==0.13.2
uvicorn==0.35.0
